+++ instrumentation/cassandra/cassandra-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/cassandra/v4_0/CassandraClientInstrumentationModule.java
@@ -6,20 +6,11 @@
 package io.opentelemetry.javaagent.instrumentation.cassandra.v4_0;
 
 import static java.util.Collections.singletonList;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.List;
-import java.util.Map;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(InstrumentationModule.class)
 public class CassandraClientInstrumentationModule extends InstrumentationModule {
@@ -30,21 +21,5 @@ public class CassandraClientInstrumentationModule extends InstrumentationModule 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
     return singletonList(new SessionBuilderInstrumentation());
-  }
-
-  public static class SessionBuilderInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      // Note: Cassandra has a large driver and we instrument single class in it.
-      // The rest is ignored in AdditionalLibraryIgnoresMatcher
-      return named("com.datastax.oss.driver.api.core.session.SessionBuilder");
-    }
-
-    @Override
-    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-      return singletonMap(
-          isMethod().and(isPublic()).and(named("buildAsync")).and(takesArguments(0)),
-          CassandraClientAdvice.class.getName());
-    }
   }
 }
